Made a few edits to the description file. Created html version
Looks good Daiva! I changed the filename to be .md, just cause that is
easier to convert things to other file formats (doc, docx, html, etc).
I editted the number of workshops from 4 to several, as I would prefer
it to be ambiguous/open-ended.  You can check out the differences in
the file by doing:

   git diff HEAD^ HEAD description.md

I created an html version of the file so that we can just copy and
paste the contents (opened in a web browser) into our email to send to
Louisa.